requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0
GitPython>=3.1.40
mysql-connector-python>=8.1.0
Jinja2>=3.1.2
packaging>=23.1
pywin32>=306
